requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
aiosqlite==0.17.0
aisuite==0.1.6
annotated-types==0.7.0
anthropic==0.40.0
anyio==4.6.2.post1
appnope==0.1.4
APScheduler==3.11.0
asttokens==2.4.1
asyncpraw==7.8.1
asyncprawcore==2.4.0
attrs==25.3.0
beautifulsoup4==4.13.4
CacheControl==0.14.2
cachetools==5.5.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
cryptography==44.0.0
cssselect==1.3.0
debugpy==1.8.9
decorator==5.1.1
distro==1.9.0
docopt==0.6.2
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.5
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
firebase-admin==6.6.0
frozenlist==1.5.0
fsspec==2024.10.0
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
groq==0.9.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.0
huggingface-hub==0.26.3
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.8.0
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==5.4.0
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.1.3
openai==1.55.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
pinecone==7.3.0
pinecone-plugin-assistant==1.7.0
pinecone-plugin-interface==0.0.7
pip-chill==1.0.3
pipdeptree==2.24.0
pipreqs==0.4.13
platformdirs==4.3.6
praw==7.8.1
prawcore==2.4.0
prompt_toolkit==3.0.48
propcache==0.3.0
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.0
scipy==1.15.0
selenium==4.33.0
sentence-transformers==4.0.1
setuptools==78.1.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.1
threadpoolctl==3.5.0
tiktoken==0.9.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.21.0
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2024.2
tzlocal==5.2
update-checker==0.18.0
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.32.1
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
yarg==0.1.10
yarl==1.18.3
zipp==3.21.0